client/requirements.txt
dbus-fast==2.39.3
dbus-python==1.3.2
gobject==0.1.0
pycairo==1.27.0
PyGObject==3.52.1
